refactor(capabilities): establish transport/runtime split with fs as template
